refactor(utils)!: Refactor utils into smaller modules
- Split utils.py into smaller focused modules
- Reorganize imports across the codebase
- Refactor has_directive as has_given_directive for clarity
- Reorganize util tests
- Fix broken tests
- Refactor and reuse is_*_type (graphql) functions throughout the codebase
- Refactor broad try/except block in naming cli with narrow ones
- Use only pathlib and do not mix it with os.path in schema_loader
- Remove read_file function and use path.read_text() instead
- Rename base.py -> graphql_type.py

BREAKING CHANGE: Multiple API changes:
1. Function signature changed from has_directive() to has_given_directive()
2. Import paths changed: from utils import → from utils.directive import
3. Module structure: utils.py split into utils/graphql_type.py, utils/directive.py, etc.
4. read_file function is removed from utils
