+++ tests/test_gql_inspector.py
@@ -4,7 +4,7 @@ from pathlib import Path
 
 import pytest
 
-from s2dm.exporters.utils import create_tempfile_to_composed_schema
+from s2dm.exporters.utils.schema_loader import create_tempfile_to_composed_schema
 from s2dm.tools.graphql_inspector import GraphQLInspector
 
 DATA_DIR: Path = Path(__file__).parent / "data"
